Updates example how_to with new API.
